+++ src/main/java/com/balaclavalab/redis/DumpRestoreCli.java
@@ -2,10 +2,15 @@ package com.balaclavalab.redis;
 
 import io.lettuce.core.KeyScanCursor;
 import io.lettuce.core.RedisClient;
+import io.lettuce.core.RedisURI;
 import io.lettuce.core.ScanArgs;
 import io.lettuce.core.api.StatefulRedisConnection;
+import io.lettuce.core.api.reactive.RedisKeyReactiveCommands;
 import io.lettuce.core.api.reactive.RedisReactiveCommands;
 import io.lettuce.core.api.sync.RedisCommands;
+import io.lettuce.core.api.sync.RedisKeyCommands;
+import io.lettuce.core.cluster.RedisClusterClient;
+import io.lettuce.core.cluster.api.StatefulRedisClusterConnection;
 import io.lettuce.core.codec.ByteArrayCodec;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
@@ -24,7 +29,9 @@ public class DumpRestoreCli {
     public static void main(String[] args) {
         Options options = new Options();
         options.addOption("f", "uriFrom", true, "Redis from (e.g. redis://localhost/1)");
-        options.addOption("t", "uriTo", true, "Redis to (e.g. redis://localhost/2)");
+        options.addOption("fc", "fromCluster", false, "Use cluster connection for Redis from");
+        options.addOption("t", "uriTo", true, "Redis to (e.g. redis-cluster://localhost/0)");
+        options.addOption("tc", "toCluster", false, "Use cluster connection for Redis to");
         options.addOption("m", "scanMatch", true, "Scan Match (default: *)");
         options.addOption("l", "scanLimit", true, "Scan Limit (default: 5000)");
 
@@ -35,11 +42,13 @@ public class DumpRestoreCli {
             if (commandLine.hasOption("help") || !hasOptions) {
                 printHelp(options);
             } else {
-                String uriFrom = commandLine.getOptionValue("f");
-                String uriTo = commandLine.getOptionValue("t");
+                RedisURI uriFrom = RedisURI.create(commandLine.getOptionValue("f"));
+                RedisURI uriTo = RedisURI.create(commandLine.getOptionValue("t"));
+                boolean fromCluster = commandLine.hasOption("fc");
+                boolean toCluster = commandLine.hasOption("tc");
                 String scanMatch = commandLine.getOptionValue("m", "*");
-                int scanLimit = Integer.valueOf(commandLine.getOptionValue("l", "5000"));
-                dumpRestore(uriFrom, uriTo, scanMatch, scanLimit);
+                int scanLimit = Integer.parseInt(commandLine.getOptionValue("l", "5000"));
+                dumpRestore(uriFrom, fromCluster, uriTo, toCluster, scanMatch, scanLimit);
             }
         } catch (ParseException e) {
             System.err.println("Parsing failed. Reason: " + e.getMessage());
@@ -52,20 +61,42 @@ public class DumpRestoreCli {
         formatter.printHelp("blc-redis-dump-restore", "BLC Redis dump restore utility", options, null, true);
     }
 
-    private static void dumpRestore(String uriFrom, String uriTo, String scanMatch, int scanLimit) {
+    private static void dumpRestore(
+            RedisURI uriFrom,
+            boolean fromCluster,
+            RedisURI uriTo,
+            boolean toCluster,
+            String scanMatch,
+            int scanLimit) {
         System.out.println("Starting dump and restore with following scanArgs: match=" + scanMatch + ", limit=" + scanLimit);
-        System.out.println("From Redis: " + uriFrom);
-        System.out.println("To Redis: " + uriTo);
+        System.out.println("From Redis: " + uriFrom + ", cluster enabled: " + fromCluster);
+        System.out.println("To Redis: " + uriTo + ", cluster enabled: " + toCluster);
         AtomicLong counter = new AtomicLong();
 
-        RedisClient clientFrom = RedisClient.create(uriFrom);
-        StatefulRedisConnection<byte[], byte[]> connectFrom = clientFrom.connect(new ByteArrayCodec());
-        RedisCommands<byte[], byte[]> commandsFromSync = connectFrom.sync();
-        RedisReactiveCommands<byte[], byte[]> commandsFrom = connectFrom.reactive();
+        RedisKeyCommands<byte[], byte[]> commandsFromSync;
+        RedisKeyReactiveCommands<byte[], byte[]> commandsFrom;
+        if (fromCluster) {
+            RedisClusterClient clusterClientFrom = RedisClusterClient.create(uriFrom);
+            StatefulRedisClusterConnection<byte[], byte[]> connectFrom = clusterClientFrom.connect(new ByteArrayCodec());
+            commandsFromSync = connectFrom.sync();
+            commandsFrom = connectFrom.reactive();
+        } else {
+            RedisClient clientFrom = RedisClient.create(uriFrom);
+            StatefulRedisConnection<byte[], byte[]> connectFrom = clientFrom.connect(new ByteArrayCodec());
+            commandsFromSync = connectFrom.sync();
+            commandsFrom = connectFrom.reactive();
+        }
 
-        RedisClient clientTo = RedisClient.create(uriTo);
-        StatefulRedisConnection<byte[], byte[]> connectTo = clientTo.connect(new ByteArrayCodec());
-        RedisReactiveCommands<byte[], byte[]> commandsTo = connectTo.reactive();
+        RedisKeyReactiveCommands<byte[], byte[]> commandsTo;
+        if (toCluster) {
+            RedisClusterClient clusterClientTo = RedisClusterClient.create(uriTo);
+            StatefulRedisClusterConnection<byte[], byte[]> connectTo = clusterClientTo.connect(new ByteArrayCodec());
+            commandsTo = connectTo.reactive();
+        } else {
+            RedisClient clientTo = RedisClient.create(uriTo);
+            StatefulRedisConnection<byte[], byte[]> connectTo = clientTo.connect(new ByteArrayCodec());
+            commandsTo = connectTo.reactive();
+        }
 
         ScanArgs scanArgs = ScanArgs.Builder.limit(scanLimit).match(scanMatch);
         KeyScanCursor<byte[]> scanCursor = commandsFromSync.scan(scanArgs);
@@ -81,8 +112,8 @@ public class DumpRestoreCli {
     private static Mono<byte[]> processKeys(
             AtomicLong counter,
             List<byte[]> keys,
-            RedisReactiveCommands<byte[], byte[]> commandsFrom,
-            RedisReactiveCommands<byte[], byte[]> commandsTo) {
+            RedisKeyReactiveCommands<byte[], byte[]> commandsFrom,
+            RedisKeyReactiveCommands<byte[], byte[]> commandsTo) {
         return Flux.fromIterable(keys)
                 .flatMap(key ->
                         Mono.zip(